english-grammar-agents/src/requirements.txt
fastapi==0.109.0
ipykernel==6.28.0
langchain==0.1.0
openai==0.28
python-dotenv==1.0.0
uvicorn==0.25.0
tiktoken==0.5.2
